requirements.txt
flask==1.1.1
requests==2.22.0
python-valve==0.2.1
bs4==0.0.1
selenium==3.141.0
Flask-WTF==0.14.2